+++ ptm-be/src/main/java/sys/app/ptm/model/request/ReleaseModelRequest.java
@@ -1,5 +1,6 @@
 package sys.app.ptm.model.request;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import lombok.Getter;
@@ -8,6 +9,6 @@ import lombok.Setter;
 @Getter @Setter
 public class ReleaseModelRequest {	
 	private String userId;
-	private float totalAmount;
+	private BigDecimal totalAmount;
 	private List<String> boards;
 }
